fix3

+++ model/BallCondition.cs
@@ -8,11 +8,11 @@ namespace DinoTem.model
     //pes 18, pes 17
     public class BallCondition
     {
-        private int id;
-	    private int frag;
-        private int unknown;
+        private UInt16 id;
+        private UInt16 frag;
+        private byte unknown;
 
-        public BallCondition(int id)
+        public BallCondition(UInt16 id)
         {
     	    if (id < 0)
                 throw new ArgumentException("Ball's id isn't valid: " + id);
@@ -20,30 +20,30 @@ namespace DinoTem.model
             this.id = id;
         }
 
-        public int getId()
+        public UInt16 getId()
         {
             return this.id;
         }
-    
-        public int getFrag()
+
+        public UInt16 getFrag()
         {
             return this.frag;
         }
 
-        public int getUnknown()
+        public byte getUnknown()
         {
             return this.unknown;
         }
 
-        public void setId(int id)
+        public void setId(UInt16 id)
         {
     	    if (id < 0)
                 throw new ArgumentException("Ball's id isn't valid: " + id);
     	
             this.id = id;
         }
-    
-        public void setFrag(int frag)
+
+        public void setFrag(UInt16 frag)
         {
     	    if (frag < 0)
                 throw new ArgumentException("Ball's frag isn't valid - Id ball: " + getId());
@@ -51,7 +51,7 @@ namespace DinoTem.model
             this.frag = frag;
         }
 
-        public void setUnknown(int unknown)
+        public void setUnknown(byte unknown)
         {
     	    if (unknown < 0)
                 throw new ArgumentException("Ball's byte isn't valid - Id ball: " + getId());
